requirements.txt
streamlit
google-generativeai
python-dotenv
streamlit-chat
pydantic
groq
ollama
ruff
pytest
joblib